Tweaks to makefiles and build scripts
I think I tracked down and eliminated all hard-coded variables in
the makefile and Compilers/*.mk files.  From now on I'd like to
confine all compilation-flag changes to the build script, while
leaving all makefile/.mk files as is; this will make it much easier
to keep track of computer-specific settings, and update for other
applications and/or computers down the road.