Modified nep5.h to version suitable to all variants
I altered the nep5.h file, using GK_nep5.h as a reference, to
serve as a base header file for all compilation variants.  I did
not explicitly define or undef the BESTNPZ or FEAST options, so
these can be set via the MY_CPP_FLAGS variable in the build script.