Corrected missing declaration in bestnpz.h
The predSumXXX variables were only defined for FEAST runs, but are needed
for BEST_NPZ runs as well.  Moved this variable declaration out of a FEAST
ifdef.

Also removed the -DBIOLOGY flag from build script... this is set automatically
by globaldefs, and was leading to annoying warning messages when it was overwritten.