Added new BioSink subroutine
New sinking subroutine is a modification of the exisiting
BIOSINK_1 subroutine, without the sinkOUT/sinkIN modifications
Georgina had added.  Those split values were added to restrict
sinking below a certain depth; that method is a bit buggy and
leads to exponential accumulation of mass at depth.

In this new version, the BioSink subroutine simply uses the rlimit
parameter to restrict sinking.  Also, I believe it will be easier
to simply flip the inputs to this routine to simulate rising,
rather than maintaining two nearly-identical subroutines for
sinking and rising.