Small changes for PROD3 index issues
The PROD3 compilation flag was causing some problems with FEAST.
Changed MV/NV limits (in mod_ncparam) to be a bit higher, then
undefed the PROD3 option.  I think I've tracked the underlying
cause to a mismatch in NPT between the ocean.in file and the
code, but until I verify that I'll keep these small adjustments
in the code.