Changed benthic loop to read all tracers in get_state
For some reason, the current version of get_state had been modified to
only read in the first of the benthic tracers from the initialization
file.  Not sure why this was, but I'm changing it back to hopefully
correct the reset-to-0 we were seeing in the DetBen group in restart
runs.