Fixed bug where NCaS should stay at bottom, not transfer to benthos
In previous version, NCaS and NCaO both transferred biomass to the
benthic detritus group when they sank out of the bottom of the
bottom layer (during diapause).  This corrects the NCaS behavior
to limit downward movement to the maximum of either 400m or the
bottom depth, preventing loss out of the bottom on the shelf.

Also some minor formatting changes (tabs to spaces).