Added TracerSink and TracerRise subroutines
Input and output to the BioSink routine was getting a little
messy, and obscured an error in the rlimit for rising.  I
added these two subroutines as wrapper functions for BioSink,
hopefully making the code a little easier to follow and removing
the need for a lot of copy/pasted code bits.