Added "defaults" to CM rising/sinking
If the values for RiseStartCM, RiseEndCM, SinkStartCM, and
SinkEndCM are all set to 0 (which is the default if they aren't
included in the input parameter file), then I assume that the
user wants the old behavior, where CM dates are 30 days after
the NC dates.

If all dates were actually set to day 0, this has the effect of
turning off diapause.  The same can be accomplished by setting
all dates to any matching value, so reserving all 0s as a special
case shouldn't limit any use cases.