My OCD finally won... updated indentation of all comments
My personal preference in code is to always keep the indentation
of comments in line with the code it is accompanying.  This code
previously kept most comments aligned hard left, which made seeing
the code structure more difficult.  It's been bugging me for months,
and I finally gave in and changed it all.

No changes to any actual code was made.