WIP: Beginning bestnpz code overhaul, added biological tracer array setup
This commit begins my overhaul of the BESTNPZ code.  The goal of
this rewrite is to make long-term maintenance of this model easier
by making the code more human-accessible.  My plan is to:

- organize the code by process, not critter.  This will make it
  easier to compare and contrast different critter formulations
  (since, e.g. feeding by MZL will be right next to feeding by Cop).
  It will also make it easier to change process equations down the
  road.

- Save all intermediate flux variables.  In addition to making
  code more legible, this will increase flexibility in outputing
  various fluxes (e.g. net production, fluxes between groups, etc.).
  Also, calculating total rate of change all at once, rather than
  one process at a time, may reduce the need for some of the
  zero-traps in the code.

- Eliminate old CPP flags, and fully document those that remain.

I do not plan to change any of the underlying math in the code
during this rewrite, with the exception of fixing any mistakes I
may encounter.

This particular commit starts by rearranging all biological tracer
variables into a consistent format, regardless of whether they
derive from the pelagic, benthic, or ice arrays.