WIP: overhauled grazing and predation
All grazing and predation (and accompanying egestion) moved to
a single section.  I also added new BasMetXXX variables to hold the
basal metabolic rate values that will be needed for respiration
calculations later on; these rates are lowered when food is scarce,
so calculating them here alongside the other food calculations
is easiest.