WIP: overhauled ice model
This section is the one where I'm still not convinced Georgina and
I are on the same page.  My interpretation of the convective
water/ice exchange, based on the Jin et al. 2006 paper, does not
seem to match her code.  I've gone over the math with pen and paper
several times, and am fairly convinced I'm doing everything correctly
in this version, but we may still need to revisit this.

We also differ in that she does the applies changes in state variables
due to adding or losing physical ice at the end, with the rest of the
ice biology calculations, while I take care of that at the beginning
when extracting state variables into their new temporary arrays.