Clarifications to BioSink, TracerSink, TracerRise subroutines
Small change to zlimit behavior in BioSink.  Rather than using
zlimit > 0 to indicate to limit no sinking, I now check to see
if zlimit is outside the range of z_w.  This accounts for the
possiblility of positive z_w values within the water column (due
to positive free surface values).  It also makes things clearer
when calling this routine in flipped-for-rising mode (TracerRise).

Corrected the comments describing inputs for the 3 subroutines,
which had gotten a bit behind the code revisions.