Overhauled sinking and rising code
This update cleans up the sinking and rising code.  That code had
been implemented in a hurried manner when trying to correct the old
biomass conservation bug in copepod diapause.  At the time I had
attempted to switch to the zlimit sinking forumalation used by the
GOANPZ code, without realizing that a) the zlimit bit hadn't been
fully implemented in the old BIOSINK_1 code, and therefore didn't
actually halt sinking at the desired depth, and b) even if the code
had been implemented identically to the GOANPZ code, that formulation
encounters problems when high sinking speeds are coupled with thin
layers (biomass concentration can go negative in that case).  So
although I like the idea behind the GOANPZ implementation that
allows critters to slow their speed when approaching the depth limit,
for now I've reverted to Georgina's method of simply adding a hard
barrier below the specified sinking depth limit.

I added the BioVert wrapper function, replacing the old TracerSink
and TracerRise wrappers.  Now one can simply call a single function
for vertical movement, and the sign of the wBio input indicates
whether movement is upward (positive) or downward (negative), and
whether the zlimit barrier acts as a ceiling or floor.