WIP: overhauled tracer reassignment
Modified portion of code where tracer variables are copied from
the scratch Bio2d and Bio3d arrays back to the main ROMS tracer
arrays.  For now I am not adding any zero traps or using the
Bio_bak subtraction method that accounts for zero traps on the
initial copy to the scratch arrays... I still need to test to see
if those sorts of manipulations will be necessary.