WIP: cleaned up FEAST portion of BESTNPZ
Just before the feast_step include statement, there was a block
of code calling several of the ROMS physics routines.  This sort
of code is not found in the typical biological module code, and
discussion with Kerim and Al suggests that it was added because
the FEAST code prescribes its own advection of fish (and therefore
requires some updates to the values along the ghost points that
join the computation tiles).  I'd like to revisit this code later
and see if we can come up with a more robust, less kludgy way of
incorporating the FEAST movement with the rest of the ROMS code;
right now this seems to be setting up problems for long-term
maintenance, since it requires cutting-and-pasting of the ROMS
tile-edge caculation code.  For now I simply wrapped the entire
block in the FEAST CPP flag, since it shouldn't be needed when
running a no-FEAST version of BEST_NPZ.