Corrected benthic feeding bug
Corrected a typo where the fraction used to distribute benthic
feeding losses across layers was a bit incorrect (didn't consider
relative concentrations in layers).  New version seems pretty
robust in tests.

Also a few changes to input and print statements associated with
compiling for Matlab mex... these are a work in progress and will
be detailed more fully in a later commit.