Added DEPTHLIMITER option to constrain NCa and Eup location
I've now wrapped the bits of code designed to force NCa and Eup
groups to stay on or off the shelf in the DEPTHLIMITER compilation
flag.  This flag turns off Eup grazing when they venture out of
their intended habitat (<200m for EupS, or >200m for EupO).  It
also kills off NCaS in water deeper than 200m when they cross the
200m depth during diapause.

I personally do not like these empirically-forced loss "processes".
This flag allows the option to add these processes and maintain
consistency with the ggibson code if desired, but does not make
that behavior the default.

Also corrected feeding preferences for EupO on detritus; this is
now controlled by the separate fpDetEupO preference, and fpDetEup
is reserved for EupS feeding.