Added euphausiid diel vertical migration, and cleaned up some TODOs
Added the EUPDIEL compilation flag.  If on, EupS performs daily
migration to depth, keeping themselves in low-light conditions.
The basics of this option are in working order, but this is still
a work in progress.

Also added some clarifying comments in place of old TODOs where
I've figured things out a bit.