Added log writes related to stationary diagnostic variables
Beginning changes to code to add new bestnpz intermediate variables
as stationary diagnostic variables (saved and available to output if
STATIONARY flag is defined).  First step here was to added write
statements to inp_par.F to print the status of the idTSvar indices
of the Hout input parameter array.