Updated modules and def_ code for new stationary array
After the consolidation of the stationary diagnostics array in
bestnpz.h, several bits of setup code needed to be updated to
reflect the new number of variables in the stationary array. I
also made some minor changes to spacing and comments in the def_
code, but that code didn't need any major changes (already set
up to loop over stationaries).

I have not touched any of the setup code related to the
now-deprecated options (PROD3, PROD2, BIOLFLUX).  However, this
code might fail if one were to turn those on, since I'm
making no effort to keep these in sync anymore.