Corrected inp_par bug related to reading of H/Sout(idTSvar) flags
Some sloppy reuse of the Lsta variable when parsing stationary
diagnostic Hout and Sout flags led to an index-out-of-bounds error
when I increased the number of diagnostics beyond the number of
tracers.  Note that this bug existed prior to my recent changes
to the stationary diagnostics code; it simply never manifested
itself due to the low number of diagnostic variables.