Added net production diagnostics, and moved i3stat stuff to new files
Net production rates for all living groups in the BEST_NPZ model are
now saved as stationary diagnostics.  This calculation is similar
to the one that used to be returned by the PROD3/PROD2 flags (but
note that the new version saves the rate at the end of a time step
rather than the total net change across that time step).

Managing all the bits of code related to the stationary diagnostic
variables was getting a bit tedious, and trying to manually keep all
the indices in sync across bestnpz.h, feast_step.h, and the variable
declaration and setup bits scattered across different modules seemed
to be asking for trouble.  I've now moved the module bits to their
own .h files, which are then #included in the proper locations.
This makes it much easier to generate those bits of code via a script
(my Matlab script romsCodeBits.m), and ensure that they remain in
sync with each other.