Removed (possibly temporarily) min function from benthic feeding
There was a minimum limit applied to benthic feeding rates that I
think may have been intended to function as a zero trap.  However,
the minimum function compared two quantities with different units.
Until I can clarify exactly what this line of code was intended to
do, I'm going to keep it commented out in favor of a version with
no minimum applied.