Added new bestnpz_swfrac_tile subroutine
In the existing code, the calculation of light attenuation appeared
to be slightly out of sync between the PAR calculations (bestnpz.h)
and swdk calculations (pre_step3d.F).  This new function has been
added to calculate the attenuation curves needed for both bits of
code.  My hope is that this centralization will make the code more
easily maintainable if we continue to make modifications to the
light attenuation scheme (as with the recent addition of the
COKELET option).