Added zero-maxes to bestnpz and light attenuation calls
I am hitting some blowup problems, so I've decided to add back the
zero-max bits to the bestnpz code.  This eliminates complete
conservation of mass.  I'd still like to look into approaches to
avoid this, but for now I want to verify that the blowup issues aren't
coming from my code specifically (note that biomass can still go
negative in the main mixing equations, though).

I also added zero-maxes to the bestnpz light attenuation code.  That's
one place where we definitely don't want any negatives, and it
doesn't mess with conservation of mass, so this zero-max should
be fine to keep in the long run.