Added freezing/melting diagnostic variables
Continuing to track down some NaNs that I think are related to ice.
I added in a diagnostic variable to track the amount of material
exchanged between the surface layer and ice skeletal layer to to
appearance or disappearance of ice.

I also made a small change to the way ice-related tracers are
extracted at the beginning of bestnpz. (Prior I checked whether ice
was present in previous and next step.  Now I skip this... shouldn't
be anything in the main tracer if ice wasn't present in the previous
step, but this might be the source of the double-counting of ice
variales on restarts).