Raised max_bio limit for biology blowup detection
The previous max_bio value was set at 10000 for all biology.  In a
few places in our domain (very shallow inlets on the GoA side), we're
getting pretty high peak PhL (around 250 g/m^2).  This is probably
too high, but due to inappropriate parameterization for that part of
the domain rather than numerical blowups.  I'm increasing the blowup
diagnostic limit to account for this.