Fixes to ice biology model
Previous problems in the ice biology were tracked to the fact that
the IceLog variable, which I had interpreted as a static logical
indicator of ice presence, was actually coded as a floating-point
active tracer.  Because of this, regions where ice formation or
melting overalpped with significant advection and diffusion held
unexpected, in-between values of IceLog.  To fix this problem, I've
now ignored the IceLog variable, and simply set the ice_status
variable based on the value of hi (ice thickness) in the current and
previous steps.

All other ice processes have been turned back on.