Reformatted mod_feast for compilation on hyak-mox
The compiler on hyak-mox is a bit pickier than the one on the older
systems, so had to make a few minor changes to mod_feast.  This mostly
involved reformatting line wraps (& placement).  Also, needed to
change the MM variable (month, used in feast calcs) to avoid decalaration
conflict with Mm (# internal eta grid points).

While I was at it, I cleaned up the indentation for the entire module.