Updated Linux-ifort.mk to work with single vs double netcdf library
The netcdf libraries on beast and cluster1 make use of a single
library file to store functions (libnetcdf.a) while on mox there
are two files (libnetcdf.a and libnetcdff.a).  This seems to be
related to earlier vs. more recent versions of netCDF.

When assigning the LIBS flags, the .mk file now checks the hostname
to see if it starts with "mox".  If it does, it assumes two files,
and if not, it assumes one file.  This is a pretty brittle fix,
but it will work for the short term at least to allow compilation
across machines without needing to constantly change the compiler
makefiles in the repository.