Corrected bug in assigning new values to benthos in bestnpz
The previous version of the code used the same method to place
the final values of Bio2D into the bt arrays as I used for the
other biological tracers.  That was wrong, because benthos are
not pre-stepped, and therefore the nnew values will not hold
relevant values yet right after restart.  The new code skips the
add-to-exisiting-nnew values method, and just assigns the Bio2d
value directly to the nnew timestep of bt.

I also removed all the debugging print statements that I used to
track this bug down.