Corrected bug in writing of first average record after restart
When calculating time-averaged values immediately following a
restart, the code included no way of accounting for a partial
record timestep.  If the NRST value is smaller than NAVG, and the
restart occurs between NAVG intervals, then the resulting averaged
value would divide by too many time steps and result in too-low
artifact for that first time record.  Added a few changes to
set_avg.F to account for this situation.