Added watts2photons parameter, and switched PAR to use W m^-2
The code previously calculated PAR in units of E/m^2/d, with hard-coded
conversions back to W m^-2 for some processes.  This change switched
things up, adding a parameter for the conversion, and keeping PAR
in the more commonly-used W m^-2 (with the watts2photons parameter
used when chl-based units are needed).  This change doesn't actually
change any of the calculations, just makes the code a bit less confusing
in prep for adding a new light limitation option.

Also deleted the KODIAK_IRAD bit of code, which was deprecated and
incorrect and just taking up space.