Added new CalcLinearCapped function
This code was repeated in a several places in the code, including
initial conditions and nudging.  It sets a depth profile to a
certain value at two points, with a linear transition between them
and constant values above and below.  Added here as a function just
to keep the code a little more reader-friendly and uncluttered.