Fixing bug dealing with BESTNPZ diagnostics in average files
The code in set_avg for accumulating values related to diagnostics
was poached from the goanpz code.  However, GOANPZ intentionally
calculates daily sums of production values (the only diagnostic
in that model), and therefore made some adjustments to the set_avg
code that was not appropriate for the bestnpz variables.  The result
was that diagnostic values in the averages code was off by a factor
related to the averaging time step.

I believe I've now fixed that mistake.