Ported mean age tracer from old age branch to latest code
This code adds an option (activated by the MEANAGE flag) that adds a
mean age tracer option.  It is based off of the AGE_MEAN option in the
main trunk, but less generic.  Also, the sources/sinks are added like
a biological tracer, since I couldn't get point sources to work away
from a land mask.